requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
prisma>=0.11.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-socketio>=5.11.0
aiohttp>=3.10.0
gunicorn>=21.2.0
pytest>=8.0.0
httpx>=0.26.0
tabulate>=0.9.0
fastapi-mail>=1.4.1
redis>=4.6.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
APScheduler>=3.10.4
python-multipart>=0.0.9
